sinnoh added

+++ rename_jfif.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 for f in lunar/*.jfif; do
-  mv "$f" "${f%.jfif}.png"
+  [ -f "$f" ] && mv "$f" "${f%.jfif}.png"
+done
+for f in lunar/*.jpg; do
+  [ -f "$f" ] && mv "$f" "${f%.jpg}.png"
 done
 echo "Done."
